refactor: mailService 리팩토링 (#117)
