--- .github/workflows/backend-cd.yml
@@ -1,41 +1,0 @@
-name: Melly Backend CD
-on:
-  push:
-    branches: [ main, develop ]
-
-jobs:
-  deploy:
-    runs-on: ubuntu-latest
-    steps:
-      - name: 저장소 Checkout
-        uses: actions/checkout@v3
-        with:
-          token: ${{secrets.ACTION_TOKEN}}
-          submodules: true
-
-      - name: 스프링부트 애플리케이션 빌드
-        run: ./gradlew build -x test
-
-      - name: 도커 이미지 빌드
-        run: docker build -t sjmin/melly .
-
-      - name: Docker Hub 로그인
-        uses: docker/login-action@v2
-        with:
-          username: ${{ secrets.DOCKERHUB_USERNAME }}
-          password: ${{ secrets.DOCKERHUB_TOKEN }}
-
-      - name: Docker Hub 퍼블리시
-        run: docker push sjmin/melly
-
-      - name: Deploy to Dev
-        uses: appleboy/ssh-action@master
-        with:
-          username: ubuntu
-          host: ${{ secrets.EC2_HOST }}
-          key: ${{ secrets.EC2_PRIVATE_KEY }}
-          script: |
-            sudo docker pull ${{ secrets.DOCKER_USERNAME }}/${{ secrets.DOCKER_REPO }}
-            chmod 777 deploy.sh
-            sh deploy.sh
-            docker image prune -f
