+++ .github/workflows/backend-ci.yml
@@ -12,10 +12,10 @@ jobs:
       - name: checkout code
         uses: actions/checkout@v2
 
-      - name: Set up JDK 11
+      - name: Set up JDK 17
         uses: actions/setup-java@v1
         with:
-          java-version: 11
+          java-version: 17
 
       - name: Gradle Caching
         uses: actions/cache@v3
