+++ core/core-api/src/main/java/cmc/mellyserver/common/event/SignupCompletedEventHandler.java
@@ -4,7 +4,7 @@ package cmc.mellyserver.common.event;
 import cmc.mellyserver.dbcore.user.User;
 import cmc.mellyserver.dbcore.user.UserRepository;
 import cmc.mellyserver.domain.comment.event.SignupCompletedEvent;
-import cmc.mellyserver.mail.EmailSendService;
+import cmc.mellyserver.mail.EmailService;
 import cmc.mellyserver.support.exception.BusinessException;
 import cmc.mellyserver.support.exception.ErrorCode;
 import lombok.RequiredArgsConstructor;
@@ -23,13 +23,13 @@ public class SignupCompletedEventHandler {
 
     private final UserRepository userRepository;
 
-    private final EmailSendService emailSendService;
+    private final EmailService emailService;
 
     @Async
     @TransactionalEventListener(phase = TransactionPhase.AFTER_COMMIT)
     public void signupEvent(SignupCompletedEvent event) {
 
         User user = userRepository.findById(event.getUserId()).orElseThrow(() -> new BusinessException(ErrorCode.USER_NOT_FOUND));
-        emailSendService.sendMail(SIGNUP_CELEBRATION_MAIL, user.getNickname(), user.getEmail());
+        emailService.send(SIGNUP_CELEBRATION_MAIL, user.getNickname(), user.getEmail());
     }
 }
