+++ core/core-api/src/test/java/cmc/mellyserver/common/mock/MockEmailClient.java
@@ -1,6 +1,7 @@
 package cmc.mellyserver.common.mock;
 
-import cmc.mellyserver.mail.EmailSendService;
+
+import cmc.mellyserver.mail.EmailSendClient;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.context.annotation.Profile;
 import org.springframework.stereotype.Component;
@@ -8,11 +9,9 @@ import org.springframework.stereotype.Component;
 @Slf4j
 @Component
 @Profile("test")
-public class MockMailService implements EmailSendService {
-
-
+public class MockEmailClient implements EmailSendClient {
     @Override
     public void sendMail(String subject, String content, String... to) {
-
+        log.info("send mail by mockEmailClient");
     }
 }
