+++ infra/mail/src/main/java/cmc/mellyserver/mail/EmailSendClient.java
@@ -1,6 +1,6 @@
 package cmc.mellyserver.mail;
 
-public interface EmailSendService {
+public interface EmailSendClient {
 
     void sendMail(String subject, String content, String... to);
 }
